refactor out footer for reuse
